experiments/profile_picture_validation/requirements.txt
fastapi
uvicorn
nudenet
Pillow
python-multipart
Jinja2
